Refactor: Moved UpsertNviPeriodDialog-file to new location (#8549)
