+++ src/pages/basic-data/nvi/reporting-periods/NviPeriodsPage.tsx
@@ -7,9 +7,9 @@ import { NviStatusMultiSelect } from '../../../../components/filters/nvi/NviStat
 import { ListSkeleton } from '../../../../components/ListSkeleton';
 import { MainContentLayout } from '../../../../components/page-layouts/MainContentLayout';
 import { NviPeriod, NviPeriodReport } from '../../../../types/nvi.types';
-import { UpsertNviPeriodDialog } from '../../../basic_data/app_admin/UpsertNviPeriodDialog';
 import { CenteredPercentageControlledCell } from '../_styles/nvi-admin-table-styles';
 import { NviAdminReportingPeriodsRow } from './_components/NviAdminReportingPeriodsRow';
+import { UpsertNviPeriodDialog } from './_components/UpsertNviPeriodDialog';
 import { useFilteredNviPeriods } from './_hooks/useFilteredNviPeriods';
 
 export const NviPeriodsPage = () => {
